API naming tweaks
