+++ vpo-backend/node-engine/src/nodes/prelude.rs
@@ -1,3 +1,5 @@
+use std::borrow::Cow;
+
 pub(super) use std::any::Any;
 pub(super) use std::collections::HashMap;
 
@@ -5,7 +7,6 @@ pub(super) use resource_manager::ResourceIndex;
 
 pub(super) use crate::errors::{NodeError, NodeOk, NodeResult, NodeWarning};
 pub(super) use crate::node::{
-    midi_input, midi_output, multiple_choice, property, stream_input, stream_output, value_input, value_output,
     InitResult, Ins, Node, NodeGraphAndIo, NodeIndex, NodeInitParams, NodeIo, NodeProcessContext, NodeRow, NodeRuntime,
     NodeState, Outs, ProcessResult,
 };
@@ -22,4 +23,49 @@ pub fn float(val: f32) -> Option<Primitive> {
 
 pub fn bool(val: bool) -> Option<Primitive> {
     Some(Primitive::Boolean(val))
+}
+
+pub fn stream_input(name: &'static str) -> NodeRow {
+    NodeRow::Input(
+        Socket::Simple(Cow::Borrowed(name), SocketType::Stream, 1),
+        SocketValue::None,
+    )
+}
+
+pub fn midi_input(name: &'static str) -> NodeRow {
+    NodeRow::Input(
+        Socket::Simple(Cow::Borrowed(name), SocketType::Midi, 1),
+        SocketValue::None,
+    )
+}
+
+pub fn value_input(name: &'static str, default: Primitive) -> NodeRow {
+    NodeRow::Input(
+        Socket::Simple(Cow::Borrowed(name), SocketType::Value, 1),
+        SocketValue::Value(default),
+    )
+}
+
+pub fn stream_output(name: &'static str) -> NodeRow {
+    NodeRow::Output(Socket::Simple(Cow::Borrowed(name), SocketType::Stream, 1))
+}
+
+pub fn midi_output(name: &'static str) -> NodeRow {
+    NodeRow::Output(Socket::Simple(Cow::Borrowed(name), SocketType::Midi, 1))
+}
+
+pub fn value_output(name: &'static str) -> NodeRow {
+    NodeRow::Output(Socket::Simple(Cow::Borrowed(name), SocketType::Value, 1))
+}
+
+pub fn property(prop_id: &str, prop_type: PropertyType, prop_default: Property) -> NodeRow {
+    NodeRow::Property(prop_id.to_string(), prop_type, prop_default)
+}
+
+pub fn multiple_choice(prop_id: &str, choices: &[&str], default_choice: &str) -> NodeRow {
+    NodeRow::Property(
+        prop_id.to_string(),
+        PropertyType::MultipleChoice(choices.iter().map(|&choice| choice.to_string()).collect()),
+        Property::MultipleChoice(default_choice.to_string()),
+    )
 }
